Split out NonExclusiveChecklistItems
