+++ editor.planx.uk/src/@planx/components/Checklist/Public/components/AutoAnsweredChecklist.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 
-import { Props } from "../types";
+import { Props } from "../../types";
 
 // An auto-answered Checklist won't be seen by the user, but still leaves a breadcrumb
 export const AutoAnsweredChecklist: React.FC<
